# openapi3_0.yaml
openapi: 3.0.0
info:
  title: Lost and Found System API
  version: 1.0.0
  description: API for managing lost and found items, auctions, and user accounts.
servers:
  - url: https://api.lostandfound.com/v1
...
paths:
  /users:
    post:
      tags:
        - User Management
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid input
...
  /auctions:
    get:
      tags:
        - Auctions
      summary: Get list of auctions
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, upcoming, past] 
          required: false
          description: Filter auctions by status (active, upcoming, past)
      responses:
        '200':
          description: A list of auctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Auction'
  /auctions/history:
    get:
      summary: View auction history
      tags:
        - Auctions
      parameters:
        - in: query
          name: userId
          schema:
            type: integer # Changed to integer for user ID
          description: User ID to filter auctions participated in
      responses:
        '200':
          description: List of past auctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Auction'
  /auctions/{auctionId}/notify:
    post:
      summary: Subscribe to auction notifications
      tags:
        - Auctions
      parameters:
        - in: path
          name: auctionId
          required: true
          schema:
            type: integer # Changed to integer for auction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer 
      responses:
        '200':
          description: Successfully subscribed to notifications
        '404':
          description: Auction not found
  /auctions/{auctionId}/notify/cancel:
    post:
      summary: Cancel auction notifications
      tags:
        - Auctions
      parameters:
        - in: path
          name: auctionId
          required: true
          schema:
            type: integer 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
      responses:
        '200':
          description: Successfully cancelled notifications
        '404':
          description: Auction not found
  /auctions/{auctionId}/pay:
    post:
      summary: Process payment for an auctioned item
      tags:
        - Auctions
      parameters:
        - in: path
          name: auctionId
          required: true
          schema:
            type: integer # Changed to integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bidderId:
                  type: integer # Changed to integer
                amount:
                  type: number
      responses:
        '200':
          description: Payment processed successfully
        '404':
          description: Auction not found
...
components:
  schemas:
...
    Auction:
      type: object
      properties:
        auctionId: 
          type: integer 
        startingPrice:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        userId: 
          type: integer 
        name:
          type: string
        email:
          type: string
        password:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
security:
  - ApiKeyAuth: []